+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Release adds support for private components and integrations with third party se
 
 **UI changes**
 
+* Change to Aspen UI.
 * Change label 'certified statement' to 'reference statement'.
 * Warning Message appears at the top of home page and login page while using an Internet Explorer browser informing the user of Internet Explorer not being supported.
 * Indicate private components with lock icon.
@@ -29,6 +30,8 @@ Release adds support for private components and integrations with third party se
 * Remove icons from project listing.
 * Add Component search filter to filter results to components owned by user.
 * Add form to create system from string or URLs.
+* Change language in interface to 'system, systems' instead of 'project, projects'.
+* Navigate users to new system form page as starting point to creating new systems.
 
 **Developer changes**
 
@@ -55,6 +58,7 @@ Release adds support for private components and integrations with third party se
 * Create a new system via name given by a string in URL.
 * Add a large set of sample components (150+) generated from STIGs.
 * Detect Apple ARM platform (e.g. 'M1 chip') and use alternate backend Dockerfile with Chromium install commented out.
+* Added SystemEvent object in controls to track system events.
 
 **Bug fixes**